demo2-python/order-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
python-dotenv
